+++ lib/blackbaud-client/api/academic_year.rb
@@ -2,15 +2,16 @@ module Blackbaud
   class AcademicYear < BlackbaudObject
     attr_accessor(:description, :ea7_academic_year_id, :end_date, :school_id, :school_name, :sessions, :short_description, :start_date, :links)
 
-    def initialize(values)
-      values["sessions"].map! {|s| Blackbaud::Session.new(s)} if values["sessions"]
+    def initialize(options)
+
+      options.fetch(:values)["sessions"].map! {|s| Blackbaud::Session.new(values: s)} if options.fetch(:values)["sessions"]
 
       ["start_date", "end_date"].each do |date|
-        send("#{date}=".intern, format_date(values["#{date}"]))
-        values.delete("#{date}")
+        send("#{date}=".intern, format_date(options.fetch(:values)["#{date}"]))
+        options.fetch(:values).delete("#{date}")
       end
 
-      super(values)
+      super
     end
 
     def connection_string