+++ lib/blackbaud-client/api/attendance_by_day_record.rb
@@ -1,0 +1,5 @@
+module Blackbaud
+  class AttendanceByDayRecord < AttendanceRecord
+    attr_accessor(:ea7_attendance_by_day_id)
+  end
+end