+++ lib/blackbaud-client/api/translation_table_entry.rb
@@ -1,0 +1,5 @@
+module Blackbaud
+  class TranslationTableEntry < BlackbaudObject
+    attr_accessor(:ea7_translation_table_entry_id, :grade)
+  end
+end